cipher-service/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bson==0.5.10
certifi==2021.5.30
charset-normalizer==2.0.4
cipher-common==0.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-ApiExceptions==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-restx==0.5.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
Naked==0.1.31
pika==1.2.0
py-healthcheck==1.10.1
pycryptodome==3.10.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
